% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_matrix_dense}
\alias{read_matrix_dense}
\title{read dense matrix as an in-memory AnnData}
\usage{
read_matrix_dense(matrix, cells, genes)
}
\description{
read dense matrix as an in-memory AnnData
}
