% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_matrix_hdf5}
\alias{read_matrix_hdf5}
\title{read hdf5 csr matrix as an in-memory AnnData in Rust}
\usage{
read_matrix_hdf5(file_path, group_path)
}
\description{
read hdf5 csr matrix as an in-memory AnnData in Rust
}
