% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_matrix_SCE}
\alias{read_matrix_SCE}
\title{read CSE matrix as an in-memory AnnData in Rust}
\usage{
read_matrix_SCE(obj, slot, cells, genes)
}
\description{
read CSE matrix as an in-memory AnnData in Rust
}
